refactor: split out object store stuffs from datanode instance
